- tasks: silence gpg warnings about not giving explicit command
if we don't give a command, it says:

   gpg: WARNING: no command supplied.  Trying to guess what you mean ...

too bad we couldn't figure out what command it guessed.  when we did our
own, it does contain the output we want, but also contains other output,
namely, it contains fingerprints for all subkeys, not just the primary
public key.  even though it's the first one and we could have used grep
with -m1 probably, was not sure this is always true.  however the fpr
line does seem to always follow the pub line, so this awk should work.
if it changes, the command will just blow up so we'll know about that